Sharing a single codebase
Rather than creating two separate code paths, this application will be written to work on both iOS and Android devices. As such, the App.js file was created in the App/Containers directory as a single entry point for both platforms.

Additonally, this commit is updating the README file with instructions for running the application.